+++ payments-core/src/main/java/com/stripe/android/CreateIntentCallback.kt
@@ -9,8 +9,8 @@ import com.stripe.android.model.SetupIntent
 sealed interface AbsCreateIntentCallback
 
 /**
- * Callback to be used when you use `PaymentSheet` and intend to create the [PaymentIntent] or
- * [SetupIntent] on your server.
+ * Callback to be used when you use `PaymentSheet` and intend to create and optionally confirm the
+ * [PaymentIntent] or [SetupIntent] on your server.
  */
 @ExperimentalPaymentSheetDecouplingApi
 fun interface CreateIntentCallback : AbsCreateIntentCallback {
@@ -18,18 +18,25 @@ fun interface CreateIntentCallback : AbsCreateIntentCallback {
     /**
      * Called when the customer confirms the payment or setup.
      *
-     * Your implementation should create a [PaymentIntent] or [SetupIntent] on your server and
-     * return its client secret or an error if one occurred.
+     * Your implementation should create and optionally confirm a [PaymentIntent] or [SetupIntent]
+     * on your server and return its client secret or an error if one occurred.
      *
-     * @param paymentMethodId The ID of the [PaymentMethod] representing the customer's payment
-     * details
+     * @param paymentMethod The [PaymentMethod] representing the customer's payment details. If your
+     * server needs the payment method, send [PaymentMethod.id] to your server and have it fetch the
+     * [PaymentMethod] object. Otherwise, you can ignore this. Don't send other properties besides
+     * the ID to your server.
+     * @param shouldSavePaymentMethod This is `true` if the customer selected the
+     * "Save this payment method for future use" checkbox. Set `setup_future_usage` on the
+     * [PaymentIntent] to `off_session` if this is `true`.
      */
-    suspend fun onCreateIntent(paymentMethodId: String): CreateIntentResult
+    suspend fun onCreateIntent(
+        paymentMethod: PaymentMethod,
+        shouldSavePaymentMethod: Boolean,
+    ): CreateIntentResult
 }
 
 /**
- * Represents the result of a [CreateIntentCallback] or
- * [CreateIntentCallbackForServerSideConfirmation].
+ * Represents the result of a [CreateIntentCallback].
  */
 @ExperimentalPaymentSheetDecouplingApi
 sealed interface CreateIntentResult {
@@ -42,29 +49,4 @@ sealed interface CreateIntentResult {
         internal val cause: Exception,
         internal val displayMessage: String? = null,
     ) : CreateIntentResult
-}
-
-/**
- * Callback to be used when you use `PaymentSheet` and intend to create and confirm the
- * [PaymentIntent] or [SetupIntent] on your server.
- */
-@ExperimentalPaymentSheetDecouplingApi
-fun interface CreateIntentCallbackForServerSideConfirmation : AbsCreateIntentCallback {
-
-    /**
-     * Called when the customer confirms the payment or setup.
-     *
-     * Your implementation should create and confirm a [PaymentIntent] or [SetupIntent] on your
-     * server and return its client secret or an error if one occurred.
-     *
-     * @param paymentMethodId The ID of the [PaymentMethod] representing the customer's payment
-     * details
-     * @param shouldSavePaymentMethod This is `true` if the customer selected the
-     * "Save this payment method for future use" checkbox. Set `setup_future_usage` on the
-     * [PaymentIntent] to `off_session` if this is `true`.
-     */
-    suspend fun onCreateIntent(
-        paymentMethodId: String,
-        shouldSavePaymentMethod: Boolean,
-    ): CreateIntentResult
 }