+++ paymentsheet-example/src/main/java/com/stripe/android/paymentsheet/example/samples/ui/server_side_confirm/complete_flow/ServerSideConfirmationCompleteFlowActivity.kt
@@ -37,7 +37,7 @@ internal class ServerSideConfirmationCompleteFlowActivity : AppCompatActivity() 
 
         paymentSheet = PaymentSheet(
             activity = this,
-            createIntentCallbackForServerSideConfirmation = viewModel::createAndConfirmIntent,
+            createIntentCallback = viewModel::createAndConfirmIntent,
             paymentResultCallback = viewModel::handlePaymentSheetResult,
         )