+++ paymentsheet/src/main/java/com/stripe/android/customersheet/CustomerSheetViewModel.kt
@@ -1,0 +1,29 @@
+package com.stripe.android.customersheet
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+import androidx.lifecycle.viewmodel.CreationExtras
+import com.stripe.android.customersheet.injection.CustomerSessionScope
+import com.stripe.android.ui.core.forms.resources.LpmRepository
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
+import javax.inject.Inject
+
+@OptIn(ExperimentalCustomerSheetApi::class)
+@CustomerSessionScope
+@Suppress("unused")
+internal class CustomerSheetViewModel @Inject constructor(
+    private val customerAdapter: CustomerAdapter,
+    private val lpmRepository: LpmRepository,
+) : ViewModel() {
+
+    private val _viewState = MutableStateFlow<CustomerSheetViewState>(CustomerSheetViewState.Loading)
+    val viewState: StateFlow<CustomerSheetViewState> = _viewState
+
+    object Factory : ViewModelProvider.Factory {
+        override fun <T : ViewModel> create(modelClass: Class<T>, extras: CreationExtras): T {
+            @Suppress("UNCHECKED_CAST")
+            return CustomerSessionViewModel.component.customerSheetViewModel as T
+        }
+    }
+}