+++ paymentsheet/src/test/java/com/stripe/android/paymentsheet/analytics/PaymentSheetEventTest.kt
@@ -21,7 +21,6 @@ class PaymentSheetEventTest {
             mode = EventReporter.Mode.Complete,
             configuration = PaymentSheetFixtures.CONFIG_CUSTOMER_WITH_GOOGLEPAY,
             isDecoupled = false,
-            isServerSideConfirmation = false,
         )
         assertThat(
             event.eventName
@@ -39,7 +38,6 @@ class PaymentSheetEventTest {
             mode = EventReporter.Mode.Complete,
             configuration = PaymentSheetFixtures.CONFIG_MINIMUM,
             isDecoupled = false,
-            isServerSideConfirmation = false,
         )
         assertThat(
             event.eventName
@@ -399,14 +397,12 @@ class PaymentSheetEventTest {
             "allows_delayed_payment_methods" to false,
             "appearance" to expectedAppearance,
             "billing_details_collection_configuration" to expectedBillingDetailsCollection,
-            "is_server_side_confirmation" to false,
         )
         assertThat(
             PaymentSheetEvent.Init(
                 mode = EventReporter.Mode.Complete,
                 configuration = PaymentSheetFixtures.CONFIG_MINIMUM,
                 isDecoupled = false,
-                isServerSideConfirmation = false,
             ).additionalParams
         ).isEqualTo(
             mapOf(
@@ -451,14 +447,12 @@ class PaymentSheetEventTest {
             "allows_delayed_payment_methods" to true,
             "appearance" to expectedAppearance,
             "billing_details_collection_configuration" to expectedBillingDetailsCollection,
-            "is_server_side_confirmation" to false,
         )
         assertThat(
             PaymentSheetEvent.Init(
                 mode = EventReporter.Mode.Complete,
                 configuration = PaymentSheetFixtures.CONFIG_WITH_EVERYTHING,
                 isDecoupled = false,
-                isServerSideConfirmation = false,
             ).additionalParams
         ).isEqualTo(
             mapOf(